fix: Register plugin BEFORE super.onCreate()
In Capacitor 7, registerPlugin() uses bridgeBuilder.addPlugin() which
adds to a list BEFORE the bridge is created. But super.onCreate() calls
load() which creates the bridge.

So we must call registerPlugin() BEFORE super.onCreate(), otherwise
the plugin is never added to the bridge and we get:
'unable to find plugin: ZXingScanner'